seo(content): /about 정본 엔티티 페이지 추가, /imhansol 공유 컴포넌트로 정리
- src/app/about/page.tsx: 동명이인과 구별되는 임한솔(Hansol Lim) 정본 소개
  페이지 신설(EntityProfilePage 공유, canonical/OG/Twitter 메타 포함)
- /imhansol 페이지를 공유 컴포넌트 기반으로 슬림화
- sitemap·layout·EntityProfilePage 관련 카피/메타 정리
